feat: Add ProductGrid component for displaying products in a grid
The code changes include adding a new ProductGrid component in the src/components/products/product-grid directory. This component is used to display a grid of products, utilizing the Product interface from the src/interfaces directory. It receives an array of products as a prop and renders each product's title. This component is imported and used in the Home page of the shop.