feat: Add revalidate constant for page caching
Add the revalidate constant with a value of 60 to the page.tsx files in the (shop)/gender/[gender] and (shop) directories. This constant is used for page caching and determines how often the page should be revalidated. By setting it to 60, the page will be revalidated every 60 seconds. This improves performance by reducing unnecessary server requests and ensures that the page content stays up to date.