feat: Remove sleep function from getStockBySlug action
The code changes remove the sleep function from the getStockBySlug action in the src/actions/product/get-stock-by-slug.ts file. The sleep function was causing a delay of 1.5 seconds, which is unnecessary for retrieving the product stock by slug. Removing the sleep function improves the performance of the action and ensures faster response times.