feat: Add getCountries action to retrieve a list of countries
This commit adds the getCountries action to the country module. The action retrieves a list of countries from the database using the Prisma ORM and returns them in ascending order by name. This addition enhances the functionality of the application by providing a way to fetch and display a list of countries.