feat: Add change-user-role action to modify user role
This commit adds the change-user-role action to the user module. The action allows for modifying the role of a user. It checks if the current user is an admin before updating the user's role in the database. Additionally, it triggers a revalidation of the '/admin/users' path to reflect the updated user role. This addition enhances the functionality of the application by providing a way for admins to change user roles.